replay viewer: add 1/2x playback speed; Space already pauses the single shipped page (#1)
The replay transport gains a 0.5x speed: command '5' selects a replay-only
half speed (ReplayHalfSpeedIndex) that advances one recorded frame every
other presentation frame (halfPhase parity in advance()), the chrome grows
a 0.5x chip, and the frame JSON's sp field carries 0.5 while it is active
(ChromeInput.speed is a float now). The wire constants ride 0.5 ahead of
the engine's integer PlaybackSpeeds, and the CI renderer fixture's WIRE
literal matches. chrome_common keeps its deliberate window.CTF_WIRE lookup
(this game emits that exact global and the Dockerfile pins the line); the
provenance test pins the patched chrome bytes. The only shipped viewer
page (index.html from client/replay_broadcast.html) already binds Space to
play/pause, so no shell forwarding was needed.

+++ src/chemistry/broadcast.nim
@@ -41,7 +41,7 @@ type
     skipLulls*: bool
     fastForwarding*: bool
     transportEnabled*: bool
-    speed*: int
+    speed*: float
     events*: JsonNode
     beats*: JsonNode
     lead*: JsonNode
